+++ src/Libraries/SmartStore.Services/DataExchange/Export/Internal/DataExporterContext.cs
@@ -25,6 +25,11 @@ namespace SmartStore.Services.DataExchange.Export.Internal
 			Projection = XmlHelper.Deserialize<ExportProjection>(request.Profile.Projection);
 			IsPreview = isPreview;
 
+			if (request.Profile.Projection.IsEmpty())
+			{
+				Projection.DescriptionMergingId = (int)ExportDescriptionMerging.Description;
+			}
+
 			FolderContent = request.Profile.GetExportFolder(true, true);
 
 			Categories = new Dictionary<int, Category>();